requirements.txt
python-telegram-bot==20.7
pytz==2024.1
python-dotenv==1.0.1
requests
beautifulsoup4
python-telegram-bot
lxml
flask